+++ src/twg2/text/tokenizer/StringParserBuilder.java
@@ -2,21 +2,18 @@ package twg2.text.tokenizer;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map.Entry;
 
 import twg2.collections.primitiveCollections.CharArrayList;
 import twg2.parser.condition.ParserCondition;
 import twg2.parser.condition.text.CharParser;
 import twg2.parser.condition.text.CharParserMatchable;
-import twg2.parser.condition.text.CharParserPredicate;
-import twg2.tuple.Tuples;
 
 /**
  * @author TeamworkGuy2
  * @since 2015-6-25
  */
 public class StringParserBuilder {
-	private List<Entry<CharParserPredicate, CharParser>> parsers;
+	private List<CharParserMatchable> parsers;
 	private boolean compound;
 	private String name;
 
@@ -27,9 +24,8 @@ public class StringParserBuilder {
 	}
 
 
-	@SuppressWarnings("unchecked")
 	public CharParserFactory build() {
-		return new CharParserMatchableFactory<CharParser>(name, compound, parsers.toArray(new Entry[parsers.size()]));
+		return new CharParserMatchableFactory<CharParser>(name, compound, parsers.toArray(new CharParserMatchable[parsers.size()]));
 	}
 
 
@@ -42,7 +38,7 @@ public class StringParserBuilder {
 		var startFilter = new StringConditions.Literal(name + "-start", new String[] { start }, Inclusion.INCLUDE);
 		var endFilter = new StringConditions.End(name + "-end", new String[] { end }, includeEnd);
 		var cond = new CharConditionPipe.OptionalSuffix<StringConditions.End>(name, false, new ParserCondition[][] { { startFilter, endFilter } });
-		this.parsers.add(Tuples.of(cond.getFirstCharMatcher(), cond));
+		this.parsers.add(cond);
 		return this;
 	}
 
@@ -56,7 +52,7 @@ public class StringParserBuilder {
 		var startFilter = new StringConditions.Literal(name + "-start", new String[] { start }, Inclusion.INCLUDE);
 		var endFilter = new CharConditions.End(name + "-end", CharArrayList.of(end), includeEnd);
 		var cond = new CharConditionPipe.OptionalSuffix<CharConditions.End>(name, false, new ParserCondition[][] { { startFilter, endFilter } });
-		this.parsers.add(Tuples.of(cond.getFirstCharMatcher(), cond));
+		this.parsers.add(cond);
 		return this;
 	}
 
@@ -70,7 +66,7 @@ public class StringParserBuilder {
 		var startFilter = new CharConditions.Literal(name + "-start", CharArrayList.of(start), Inclusion.INCLUDE);
 		var endFilter = new StringConditions.End(name + "-end", new String[] { end }, includeEnd);
 		var cond = new CharConditionPipe.OptionalSuffix<StringConditions.End>(name, false, new ParserCondition[][] { { startFilter, endFilter } });
-		this.parsers.add(Tuples.of(cond.getFirstCharMatcher(), cond));
+		this.parsers.add(cond);
 		return this;
 	}
 
@@ -84,7 +80,7 @@ public class StringParserBuilder {
 		var startFilter = new CharConditions.Literal(name + "-start", CharArrayList.of(start), Inclusion.INCLUDE);
 		var endFilter = new CharConditions.End(name + "-end", CharArrayList.of(end), includeEnd);
 		var cond = new CharConditionPipe.OptionalSuffix<CharConditions.End>(name, false, new ParserCondition[][] { { startFilter, endFilter } });
-		this.parsers.add(Tuples.of(cond.getFirstCharMatcher(), cond));
+		this.parsers.add(cond);
 		return this;
 	}
 
@@ -98,7 +94,7 @@ public class StringParserBuilder {
 		var startFilter = new CharConditions.Literal(name + "-start", CharArrayList.of(start), Inclusion.INCLUDE);
 		var endFilter = new CharConditions.EndNotPrecededBy(name + "-end", CharArrayList.of(end), includeEnd, CharArrayList.of(notPreced));
 		var cond = new CharConditionPipe.OptionalSuffix<CharConditions.EndNotPrecededBy>(name, false, new ParserCondition[][] { { startFilter, endFilter } });
-		this.parsers.add(Tuples.of(cond.getFirstCharMatcher(), cond));
+		this.parsers.add(cond);
 		return this;
 	}
 
@@ -108,7 +104,7 @@ public class StringParserBuilder {
 	 */
 	public StringParserBuilder addStringLiteralMarker(String name, String str) {
 		var cond = new StringConditions.Literal(name, new String[] { str }, Inclusion.INCLUDE);
-		this.parsers.add(Tuples.of(cond.getFirstCharMatcher(), cond));
+		this.parsers.add(cond);
 		return this;
 	}
 
@@ -118,7 +114,7 @@ public class StringParserBuilder {
 	 */
 	public StringParserBuilder addStringLiteralMarker(String name, String... strs) {
 		var cond = new StringConditions.Literal(name, strs, Inclusion.INCLUDE);
-		this.parsers.add(Tuples.of(cond.getFirstCharMatcher(), cond));
+		this.parsers.add(cond);
 		return this;
 	}
 
@@ -128,26 +124,20 @@ public class StringParserBuilder {
 	 */
 	public StringParserBuilder addCharLiteralMarker(String name, char... chars) {
 		var cond = new CharConditions.Literal(name, CharArrayList.of(chars), Inclusion.INCLUDE);
-		this.parsers.add(Tuples.of(cond.getFirstCharMatcher(), cond));
+		this.parsers.add(cond);
 		return this;
 	}
 
 
 	public StringParserBuilder addCharMatcher(String name, char[] chars) {
 		var cond = new CharConditions.Contains(name, CharArrayList.of(chars), Inclusion.INCLUDE);
-		this.parsers.add(Tuples.of(cond.getFirstCharMatcher(), cond));
+		this.parsers.add(cond);
 		return this;
 	}
 
 
 	public StringParserBuilder addConditionMatcher(CharParserMatchable conditionFilter) {
-		this.parsers.add(Tuples.of(conditionFilter.getFirstCharMatcher(), conditionFilter));
-		return this;
-	}
-
-
-	public StringParserBuilder addConditionMatcher(CharParser condition, CharParserMatchable conditionFilter) {
-		this.parsers.add(Tuples.of(conditionFilter.getFirstCharMatcher(), condition));
+		this.parsers.add(conditionFilter);
 		return this;
 	}